requirements.txt
fastapi==0.115.4
uvicorn==0.30.6
pydantic==2.9.2
motor==3.6.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
starlette==0.40.0
